Feat: Inventory 컴포넌트 추가
